requirements.txt
coverage>=5.1
mox>=0.5.3
nose>=1.3.7
tika==1.24
PyYAML==5.1.1
python-stdnum==1.11
langdetect==1.0.7
fuzzywuzzy==0.17.0
python-Levenshtein==0.12.2
spacy==2.3.0
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-2.3.0/es_core_news_sm-2.3.0.tar.gz
https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-2.3.0/xx_ent_wiki_sm-2.3.0.tar.gz

